project/final/hetznerfuncs/requirements.txt
fastapi==0.110.1
hcloud==1.35.0
python-dotenv==1.0.1
